requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.15.1
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-telegram-bot==13.12
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
typing==3.7.4.3
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
Werkzeug==2.2.2
whitenoise==6.2.0
